+++ Assets/SteamVR_Input/ActionSetClasses/SteamVR_Input_ActionSet_Menu.cs
@@ -96,5 +96,13 @@ namespace Valve.VR
                 return SteamVR_Actions.menu_MenuNestedToggle;
             }
         }
+        
+        public virtual SteamVR_Action_Vector2 MenuNavigateJoystickPosition
+        {
+            get
+            {
+                return SteamVR_Actions.menu_MenuNavigateJoystickPosition;
+            }
+        }
     }
 }